poor navigation design... but it does work
To change the window size in some sepcific pages, I used "Callback" to
pass the window ref. It is really not a good idea. But I found no way
else.

And there are lots of parts in my code are duplicated... It is really
bad.

I will implement fade-in & fade-out appearance later.